Update and rename poem.txt to week-01-murray-poem.md
Hi Maria! Your poem is great! 
There are a few issues which can be easily fixed which would keep this file from rendering (displaying) correctly in GitHub.

The extension for this file should be '.md' which is a Markdown file. Without that, GitHub won't know what kind of file this is and won't be able to render it correctly.
The images need an exclamation '!' and then a description between brackets '[description]' and then the URL (http://www.website.com/image.jpg)
Altogether it looks like '![description](http://www.website.com/image.jpg)

I tried to fix the image you had but it looks like it's no longer available at the URL that was being used. You can add a different image to make sure it all works.

For the plain text links to youtube, you'll just want the description and the url to be right next to each other with no spaces. I removed the space separating them in my proposed changes.

Let me know if you have questions!